Save base buffers when applying changes from their branches
